refactored parse
